+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## 0.2.0 - 2023-05-11
+
+- update README links
+- rename UnexpectedMetadataChange to UnsupportedMetadataChangeError
+
 ## 0.1.1 - 2023-05-10
 
-* Improved documentation
+- Improved documentation
 
 ## 0.1.0 - 2023-05-09
 
-* Initial release
+- Initial release